+++ traffic_ops_ort/atstccfg/config/config.go
@@ -250,6 +250,9 @@ type TOData struct {
 	// ServerParams must be all Parameters on the Profile of the current server. Must not include other Parameters.
 	ServerParams []tc.Parameter
 
+	// RemapConfigParams must be all Parameters with the ConfigFile atscfg.RemapConfigParameterConfigFile.
+	RemapConfigParams []tc.Parameter
+
 	// CacheKeyParams must be all Parameters with the ConfigFile atscfg.CacheKeyParameterConfigFile.
 	CacheKeyParams []tc.Parameter