+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [5.0.0] - 2020-10-20
 ### Added
+- Traffic Ops Ort: Add support for raw remap text override via DS profile parameters override.edge/remap.config and override.mid/remap.config
 - Traffic Ops Ort: Disabled ntpd verification (ntpd is deprecated in CentOS)
 - Traffic Ops Ort: Adds a transliteration of the traffic_ops_ort.pl perl script to the go language. See traffic_ops_ort/t3c/README.md.
 - Traffic Ops API v3